refactor: deduplicate parseRiskAreas and avoid double WIP computation
- Extract parseRiskAreas to classify.ts as shared utility (was duplicated
  in pr-drill-down-sheet.tsx and size-badge-popover.tsx)
- Export computeWipCounts so clientLoader can call it once and pass the
  result to both aggregateWipCycle and computeWipLabels
